Rename 'expectations' to 'use_fixture_repo'
