+++ spec/lib/bookbinder/commands/generate_pdf_spec.rb
@@ -1,7 +1,7 @@
 require 'yaml'
 require_relative '../../../../lib/bookbinder/commands/generate_pdf'
 require_relative '../../../../lib/bookbinder/sitemap_generator'
-require_relative '../../../helpers/expectations'
+require_relative '../../../helpers/use_fixture_repo'
 require_relative '../../../helpers/tmp_dirs'
 require_relative '../../../helpers/nil_logger'
